refactor: ユーザ登録に使うスキーマファイルの名前の変更
