+++ src/features/admin/user/CreateUserForm.tsx
@@ -2,7 +2,7 @@ import { TextInput, MultiSelect, Select } from '@mantine/core';
 import { Button } from '@mantine/core';
 import { useForm, zodResolver } from '@mantine/form';
 import { useEffect } from 'react';
-import { schema } from './roles/shema';
+import { schema } from './roles/userShema';
 import { useSelectBeacons } from '@/features/admin/user/hooks/beaconSelector';
 import { useRoles } from '@/features/admin/user/hooks/editingUserState';
 import { useSelectTags } from '@/features/admin/user/hooks/tagSelector';
